requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.0
SQLAlchemy==2.0.16
Werkzeug==3.0.0
Flask-Login==0.6.3
WTForms-3.1.2